refactor(frontend): 简化数据层，将 useWikiArticles 逻辑内联到 App.tsx\n\n- 直接在 App.tsx 使用 useInfiniteQuery 与当前语言执行抓取\n- 使用 motion/react useScroll 在滚动到 80% 时加载下一页（含滞回）\n- 移除 hooks/useWikiArticles.ts 与旧的 IntersectionObserver 逻辑\n- useMemo/useCallback 稳定依赖，修复 lint 警告
