+++ frontend/package.json
@@ -24,6 +24,7 @@
     "@tanstack/react-query": "5.84.2",
     "@vercel/analytics": "^1.4.1",
     "lucide-react": "^0.474.0",
+    "motion": "11.18.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "sharp": "^0.34.3",
